pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "code-review-agent"
version = "0.1.0"
description = "AI-driven code review agent for CI/CD pipelines"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = []
keywords = ["code-review", "ai", "ci-cd", "adk"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "google-adk>=1.25.0",
    "httpx>=0.28.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "typer>=0.12",
    "litellm>=1.50",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=4.0",
    "ruff>=0.6",
    "respx>=0.21",
]
observability = [
    "prometheus-client>=0.19",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp-proto-http>=1.20",
]

[project.scripts]
code-review = "code_review.__main__:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "e2e: E2E tests (require Docker + Gitea, run with RUN_E2E=1)",
    "respx: HTTP mocking with respx",
]
filterwarnings = [
    # Known third-party deprecations / noisy warnings; safe to ignore for now.
    "ignore:.*PytestUnknownMarkWarning:pytest.PytestUnknownMarkWarning",
    "ignore:Inheritance class AiohttpClientSession from ClientSession is discouraged:DeprecationWarning:google.genai._api_client",
    "ignore:.*PytestUnhandledThreadExceptionWarning:pytest.PytestUnhandledThreadExceptionWarning",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "C4", "UP"]
